requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
simpy==4.1.1
six==1.17.0
